fix: some edge cases on tests

+++ src/contracts/core/common/Events.sol
@@ -19,7 +19,7 @@ library DiamondEvent {
     /**
      * @dev Triggered when the contract has been initialized or reinitialized.
      */
-    event Initialized(address indexed operator, uint8 version);
+    event Initialized(address indexed operator, uint96 version);
 }
 
 /**
